add some chages

+++ app/Http/Controllers/ProductsController.php
@@ -4,10 +4,14 @@ namespace App\Http\Controllers;
 
 use App\Models\Category;
 use App\Models\Color;
+use App\Models\Collection;
+use App\Models\Currency;
+use App\Models\ProductSize;
 use App\Models\Image;
 use App\Models\ModelEntry;
 use App\Models\Product;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\File;
 use Illuminate\Support\Str;
 
@@ -29,6 +33,9 @@ class ProductsController extends Controller
         $categories = Category::all();
         $models = ModelEntry::all();
         $colors = Color::all();
+        $collections = Collection::all();
+        $currencies = Currency::all();
+        $sizes = ProductSize::all();
 
         return view(
             'admin.products.edit'
@@ -37,51 +44,80 @@ class ProductsController extends Controller
                 , 'categories' => $categories
                 , 'models' => $models
                 , 'colors' => $colors
+                , 'collections' => $collections
+                , 'currencies' => $currencies
+                , 'sizes' => $sizes
             ]);
     }
-    public function update(Request $request){
-
+    public function update(Request $request)
+    {
+        // Валидация входных данных
         $request->validate([
-            'id' => 'required',
+            'id' => 'required|exists:products,id',
             'name' => 'required',
-            'description' => 'required',
             'category_id' => 'required',
             'model_id' => 'required',
             'color_id' => 'required',
-            'image' => 'required|file',
+            'collection_id' => 'required',
+            'size_id' => 'required',
+            'composition' => 'required',
+            'min_quantity' => 'required',
+            'price' => 'required',
+            'currency_id' => 'required',
+            'description' => 'required',
+            'image' => 'sometimes|file',
         ]);
 
-        $product = Product::find($request->get('id'));
+        // Найти продукт
+        $product = Product::find($request->id);
 
         if ($product) {
+            // Используем транзакцию, чтобы гарантировать целостность операции
+            DB::transaction(function () use ($request, $product) {
+                // Если новое изображение загружено
+                if ($request->hasFile('image')) {
+                    // Удалить старое изображение
+                    $filePath = public_path('uploads') . '/' . $product->image_url;
+                    if (File::exists($filePath)) {
+                        File::delete($filePath);
+                    }
 
-            $filePath = public_path('uploads'). '/' . $product->image_url;
+                    // Сохранить новое изображение
+                    $file = $request->file('image');
+                    $fileName = Str::random(20) . '.' . $file->getClientOriginalExtension();
+                    $file->move(public_path('uploads'), $fileName);
 
-            if (File::exists($filePath)) {
-                File::delete($filePath);
-            }
+                    // Обновить поле image_url
+                    $product->image_url = $fileName;
+                }
 
-            $file = $request->file('image');
-            $fileName = Str::random(20) . '.' . $file->getClientOriginalExtension();
+                // Обновить остальные поля продукта
+                $product->update([
+                    'name' => $request->name,
+                    'category_id' => $request->category_id,
+                    'model_id' => $request->model_id,
+                    'color_id' => $request->color_id,
+                    'size_id' => $request->size_id,
+                    'collection_id' => $request->collection_id,
+                    'composition' => $request->composition,
+                    'min_quantity' => $request->min_quantity,
+                    'price' => $request->price,
+                    'currency_id' => $request->currency_id,
+                    'description' => $request->description,
+                    'new' => $request->has('new') ? true : false,  // Если параметр new есть, то true, иначе false
+                    'pop' => $request->has('pop') ? true : false,  // То же самое для pop
+                    'active' => $request->has('active') ? true : false  // И для active
+                ]);
+            });
 
-            $file->move(public_path('uploads'), $fileName);
-
-            $product->update([
-                'name' => $request->get('name'),
-                'description' => $request->get('description'),
-                'category_id' => $request->get('category_id'),
-                'model_id' => $request->get('model_id'),
-                'color_id' => $request->get('color_id'),
-                'image_url' => $fileName,
-                'new' => $request['new'] ? true : false,
-                'pop' => $request['pop'] ? true : false,
-                'active' => $request['active'] ? true : false,
-            ]);
-
-            return redirect()->route('admin-product-show', ['id' => $request->get('id')])->with('success', 'Product updated successfully');
+            // Перенаправление после успешного обновления
+            return redirect()->route('admin-product-show', ['id' => $request->id])->with('success', 'Product updated successfully');
         }
-        return redirect()->back()->with('error', 'Image not found');
+
+        // Если продукт не найден, вернуть ошибку
+        return redirect()->back()->with('error', 'Product not found');
     }
+
 
     public function delete($id){
         $product = Product::find($id);
@@ -114,6 +150,9 @@ class ProductsController extends Controller
         $categories = Category::all();
         $models = ModelEntry::all();
         $colors = Color::all();
+        $collections = Collection::all();
+        $currencies = Currency::all();
+        $sizes = ProductSize::all();
 
         return view(
             'admin.products.create'
@@ -121,16 +160,25 @@ class ProductsController extends Controller
                 'categories' => $categories
                 , 'models' => $models
                 , 'colors' => $colors
+                , 'collections' => $collections
+                , 'currencies' => $currencies
+                , 'sizes' => $sizes
             ]
         );
     }
     public function store(Request $request){
         $request->validate([
             'name' => 'required',
-            'description' => 'required',
             'category_id' => 'required',
             'model_id' => 'required',
             'color_id' => 'required',
+            'size_id' => 'required',
+            'collection_id' => 'required',
+            'composition' => 'required',
+            'min_quantity' => 'required',
+            'price' => 'required',
+            'currency_id' => 'required',
+            'description' => 'required',
             'image' => 'required|file',
         ]);
 
@@ -140,10 +188,16 @@ class ProductsController extends Controller
         $file->move(public_path('uploads'), $fileName);
 
         Product::create([
+            'name' => $request['name'],
             'category_id'=> $request['category_id'],
             'model_id' => $request['model_id'],
             'color_id' => $request['color_id'],
-            'name' => $request['name'],
+            'size_id' => $request['size_id'],
+            'collection_id' => $request['collection_id'],
+            'composition' => $request['composition'],
+            'min_quantity' => $request['min_quantity'],
+            'price' => $request['price'],
+            'currency_id' => $request['currency_id'],
             'description' => $request['description'],
             'image_url' => $fileName,
             'new' => $request['new'] ? true : false,
